% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{digits}
\alias{digits}
\title{digits}
\usage{
digits(x, n)
}
\arguments{
\item{x}{a numeric}

\item{n}{a numeric. The number of digits to round to.}
}
\description{
Allows the user to specify the exact number of digits
}
