% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperwork.R
\name{evalPlaceholder}
\alias{evalPlaceholder}
\title{evalPlaceholder}
\usage{
evalPlaceholder(path2raw_, env_ = .GlobalEnv)
}
\arguments{
\item{path2raw_}{path to raw file}
}
\description{
Eval where '&&' exists.
}
