% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{glue_bracket}
\alias{glue_bracket}
\title{glue_bracket}
\usage{
glue_bracket(x, ..., round = 2)
}
\arguments{
\item{x}{a character or numeric}

\item{...}{a function. If provided will transform effects and confidence intervals.}

\item{round}{a scalar.}
}
\description{
glue_bracket
}
