% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_change.R
\name{m_iqr}
\alias{m_iqr}
\title{m_iqr}
\usage{
m_iqr(x, round = 1, na.rm = T)
}
\arguments{
\item{x}{a character vector}

\item{round}{a numeric.}

\item{na.rm}{a bool.}
}
\description{
takes in a numeric vector, calculates a median and iqr and returns as text.
}
