% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{n_percent}
\alias{n_percent}
\title{n_percent}
\usage{
n_percent(vector, x, round = 2, na.rm = F)
}
\arguments{
\item{vector}{a vector to count within}

\item{x}{character, the object to count by}

\item{round}{scalar, the number of digits}

\item{na.rm}{bool, whether to remove NAs}
}
\value{
a character
}
\description{
Returns the number than meet a condition, and the percentage in brackets
}
