% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperwork.R
\name{newDiary}
\alias{newDiary}
\title{newDiary}
\usage{
newDiary(
  filename_ = paste0(format(Sys.Date(), "\%Y\%m\%d"), ".Rmd"),
  template_ = "~/diary.Rmd"
)
}
\description{
Copy the diary template and execute the marked section.
}
