% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{round_p}
\alias{round_p}
\title{round p}
\usage{
round_p(p, n = 2)
}
\arguments{
\item{p}{a p value}

\item{n}{a numeric. The number of digits to round to.}
}
\description{
Rounds p value to specified digits and uses less symbol if result it zero.
}
