% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_docx.R
\name{to_docx}
\alias{to_docx}
\title{to_docx}
\usage{
to_docx(table, path, table_name = NULL)
}
\arguments{
\item{table}{a dataframe or list of data.frames}

\item{path}{a string. Where you want to save the file.}

\item{table_name}{string. If provided, allows tables to be named}
}
\description{
Sends a data.frame to a word doc.
}
