% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_docx.R
\name{to_docx}
\alias{to_docx}
\title{to_docx}
\usage{
to_docx(table, path)
}
\arguments{
\item{table}{a dataframe or tibble}

\item{path}{a string. Where you want to save the file.}
}
\description{
Sends a data.frame to a word doc.
}
