% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{to_docx}
\alias{to_docx}
\title{to_docx}
\usage{
to_docx(table, path, title = NULL, note = NULL, landscape = F)
}
\arguments{
\item{table}{a dataframe or list of data.frames}

\item{path}{a string. Where you want to save the file.}

\item{title}{string. If provided, allows tables to be named}

\item{note}{a string. Allows for notes}

\item{landscape}{a bool. If true, outputs a landscape word doc.}
}
\description{
Sends a data.frame to a word doc.
}
